Refactor internal `assembly.expression` src-files
